Continue rapport

+++ autopush.fish
@@ -4,8 +4,11 @@ git pull --rebase --autostash
 while true
 	pdfinfo rapport/main.pdf | rg ^Pages: | awk '{print $2}' > pages_count
 	git add rapport/*.typ bib.yaml *.fish rapport/*.dot pages_count rapport/*.png
+	set typst_changed (git diff --exit-code rapport/*.{typ,dot,png} bib.yaml)
 	git commit --quiet -m "Continue rapport"
-	git push --quiet --force
-	echo Pushed at (date)
+	if $typst_changed
+		git push --quiet --force
+		echo Pushed at (date)
+	end
 	sleep (math "60 * 30")
 end
